tests/wpt/tests/html/canvas/offscreen/transformations/2d.transformation.setTransform.multiple.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.transformation.setTransform.multiple</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.setTransform.multiple</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.setTransform(1/2,0, 0,1/2, 0,0);
  ctx.setTransform();
  ctx.setTransform(2,0, 0,2, 0,0);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 50, 25);
  _assertPixel(canvas, 75,35, 0,255,0,255);
  t.done();

});
</script>
